backend/requirements.txt

fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
# centrifuge==4.1.0
websockets==11.0.3
python-dotenv==1.0.0
